Test Unit

+++ src/main/java/com/finalproject/automated/refactoring/tool/nom/detection/service/implementation/NOMDetectionImpl.java
@@ -13,6 +13,6 @@ public class NOMDetectionImpl implements NOMDetection {
         if(classModel.getMethodModels() == null){
             return 0L;
         }
-        return (Long) (long) classModel.getMethodModels().size();
+        return Long.valueOf(classModel.getMethodModels().size());
     }
 }
